Move modules around a bit
