+++ protostar/cairo_testing/cairo_test_cheatcode_factory.py
@@ -1,39 +1,40 @@
 from typing import List, cast
 
-from protostar.starknet.new_arch.cheatable_cairo_cached_state import (
-    CheatableCairoCachedState,
+from protostar.cheatable_starknet.cheatable_cached_state import (
+    CheatableCachedState,
 )
-from protostar.starknet.new_arch.cheaters import (
-    BlockInfoCairoCheater,
-    ContractsCairoCheater,
+from protostar.cheatable_starknet.cheaters.block_info import BlockInfoCairoCheater
+from protostar.cheatable_starknet.cheaters.contracts import ContractsCairoCheater
+
+from protostar.cheatable_starknet.cheaters import (
     CairoCheaters,
 )
-from protostar.testing.new_arch.cairo_test_execution_state import (
+from protostar.cairo_testing.cairo_test_execution_state import (
     CairoTestExecutionState,
 )
-from protostar.testing.new_arch.cheatcodes.cairo_cheatcode import CairoCheatcode
-from protostar.testing.new_arch.cheatcodes.declare_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.cairo_cheatcode import CairoCheatcode
+from protostar.cheatable_starknet.cheatcodes.declare_cairo_cheatcode import (
     DeclareCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.deploy_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.deploy_cairo_cheatcode import (
     DeployCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.deploy_contract_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.deploy_contract_cairo_cheatcode import (
     DeployContractCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.invoke_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.invoke_cairo_cheatcode import (
     InvokeCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.prepare_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.prepare_cairo_cheatcode import (
     PrepareCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.roll_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.roll_cairo_cheatcode import (
     RollCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.warp_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.warp_cairo_cheatcode import (
     WarpCairoCheatcode,
 )
-from protostar.testing.new_arch.cheatcodes.call_cairo_cheatcode import (
+from protostar.cheatable_starknet.cheatcodes.call_cairo_cheatcode import (
     CallCairoCheatcode,
 )
 
@@ -46,9 +47,7 @@ class CairoTestCheatcodeFactory:
         self.state = state
 
     def build_cheatcodes(self) -> List[CairoCheatcode]:
-        cheatable_state = cast(
-            CheatableCairoCachedState, self.state.starknet.state.state
-        )
+        cheatable_state = cast(CheatableCachedState, self.state.starknet.state.state)
         cheaters = CairoCheaters(
             block_info=BlockInfoCairoCheater(cheatable_state=cheatable_state),
             contracts=ContractsCairoCheater(cheatable_state=cheatable_state),
