+++ protostar/cheatable_starknet/cheatcodes/declare_cairo_cheatcode.py
@@ -2,7 +2,7 @@ import asyncio
 from typing import Any, Protocol
 
 from protostar.compiler import ProjectCompiler
-from protostar.testing.new_arch.cheatcodes.cairo_cheatcode import CairoCheatcode
+from protostar.cheatable_starknet.cheatcodes.cairo_cheatcode import CairoCheatcode
 from protostar.contract_types import DeclaredContract
