+++ protostar/cheatable_starknet/cheatcodes/prepare_cairo_cheatcode.py
@@ -1,12 +1,13 @@
 from typing import Any, Callable, Optional
 import asyncio
 
+from protostar.cheatable_starknet.cheatcodes.cairo_cheatcode import CairoCheatcode
+from protostar.cheatable_starknet.cheaters.contracts import ContractsCheaterException
+
 from protostar.starknet import CheatcodeException
 from protostar.starknet.data_transformer import (
     CairoOrPythonData,
 )
-from protostar.starknet.new_arch.cheaters.contracts import ContractsCheaterException
-from protostar.testing.new_arch.cheatcodes.cairo_cheatcode import CairoCheatcode
 from protostar.contract_types import DeclaredContract, PreparedContract
